refactor: sign in -> auth package
